+++ extensions/quick-search/src/utils/readtxt.ts
@@ -1,10 +1,12 @@
 import { getSelectedText, Clipboard } from "@raycast/api";
 
 export const isNotEmpty = (string: string | null | undefined): string is string => {
-  return string != null && String(string).length > 0;
+  return string != null && String(string).trim().length > 0;
 };
 
-export const readtext = () =>
-  getSelectedText()
-    .then((text) => (isNotEmpty(text) ? text : Clipboard.readText()))
-    .catch(() => undefined);
+export const readtext = (fallbackText?: string) =>
+  isNotEmpty(fallbackText)
+    ? fallbackText?.trim()
+    : getSelectedText()
+        .then((text) => (isNotEmpty(text) ? text : Clipboard.readText()))
+        .catch(() => undefined);